+++ packages/node/src/behavior/state/managed/ManagedReference.ts
@@ -28,7 +28,7 @@ type Container = Record<string | number, Val>;
  * separate ManagedReference.
  */
 export class ManagedReference implements ValReference {
-    primaryKey;
+    readonly primaryKey = "name";
     parent;
     subrefs?: Record<number | string, ValReference>;
     owner?: Val;
@@ -46,7 +46,6 @@ export class ManagedReference implements ValReference {
 
     /**
      * @param parent a reference to the container we reference
-     * @param primaryKey the preferred key for lookup
      * @param name the name (in the case of structs) or index (in case of lists)
      * @param id the lookup ID in the case of structs
      * @param assertWriteOk enforces ACLs and read-only
@@ -55,14 +54,12 @@ export class ManagedReference implements ValReference {
      */
     constructor(
         parent: ValReference<Val.Collection>,
-        primaryKey: "name" | "id",
         name: string | number,
         id: number | undefined,
         assertWriteOk: (value: Val) => void,
         clone: (container: Val) => Val,
         session: AccessControl.Session,
     ) {
-        this.primaryKey = primaryKey;
         this.parent = parent;
         this.#assertWriteOk = assertWriteOk;
         this.#clone = clone;
@@ -73,8 +70,8 @@ export class ManagedReference implements ValReference {
             path: parent.location.path.at(name),
         };
 
-        const key = primaryKey === "id" ? (id ?? name) : name;
-        const altKey = primaryKey === "id" ? (key === name ? undefined : name) : id;
+        const key = parent.primaryKey === "id" ? (id ?? name) : name;
+        const altKey = parent.primaryKey === "id" ? (key === name ? undefined : name) : id;
         this.#key = key;
         this.#altKey = altKey;